Update dashboard and backend APIs

+++ backend/app/models/attendance.py
@@ -21,7 +21,7 @@ class Attendance(Base):
     __tablename__ = "attendances"
 
     id = Column(BigInteger, primary_key=True, autoincrement=True)
-    scanned_batch_id = Column(Integer, nullable=False, index=True)
+    scanned_batch_id = Column(BigInteger, nullable=False, index=True)
     employee_id = Column(Integer, ForeignKey("employees.id"), nullable=True)
     bus_id = Column(String(10), ForeignKey("buses.bus_id"), nullable=True, index=True)
     van_id = Column(Integer, ForeignKey("vans.id"), nullable=True, index=True)
