+++ backend/app/models/employee.py
@@ -2,7 +2,7 @@
 Employee model.
 """
 
-from sqlalchemy import Column, Integer, String, Boolean, ForeignKey
+from sqlalchemy import Column, Integer, BigInteger, String, Boolean, ForeignKey
 from sqlalchemy.orm import relationship
 
 from app.core.db import Base
@@ -14,7 +14,7 @@ class Employee(Base):
     __tablename__ = "employees"
 
     id = Column(Integer, primary_key=True, autoincrement=True)
-    batch_id = Column(Integer, unique=True, nullable=False, index=True)
+    batch_id = Column(BigInteger, unique=True, nullable=False, index=True)
     name = Column(String(100), nullable=False)
     bus_id = Column(String(10), ForeignKey("buses.bus_id"), nullable=False, index=True)
     van_id = Column(Integer, ForeignKey("vans.id"), nullable=True, index=True)
